requirements.txt
absl-py==0.11.0
aiogram==2.11.2
aiohttp==3.7.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
Keras-Preprocessing==1.1.2
Markdown==3.3.3
multidict==5.0.2
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
termcolor==1.1.0
typeguard==2.10.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3

opencv-python~=4.4.0.46